Trying out base class impl
