+++ src/password-field/index.ts
@@ -1,0 +1,1 @@
+export * from './PasswordField.js';
